Added basic datetime formatting functionality
May still be buggy, but can easily be changed as timestamps are stored
internally as unix + millis, so the system is ambivalent towards the
presentation.

The system is simple, with backslashes indicating where things need to
be inserted into the template. I will expand its functionality in the
future if I can find a way to do so.