Small naming changes, to be expanded on later.
Out of time for the day, so I'm just pushing what I have. Right now
the app mockup is still a mass naming-wise, so I'm working on
relabelling the elements in the main HMTL file of the app and the
asociated stylesheets.

I will also follow this up with better organisation for my JS code,
but as I'm still in the middle of writing more code, that's of
lesser concern.