Removed id attributes in the main page.
Not relying on ids will make css precendence rules easier to manage.
The only rule that now uses ids (the html element has an id) is hiding
elements that are supposed to be invisible. For example, the hidden
attribute is not enforced by the browser, so it has to be enforced in
the css.

Using an id selector for this purpose means that this behaviour can
still be overridden fairly easily, but it also means that doing so by
accident becomes very unlikely.

This should offer the greatest amount of customisability using custom
themes.