Added parsing for pasting in a text/html document fragment
I will organise and optimise the code later. I really just want to
commit it now that it's working-ish. Previously when pasting in html
code, the structure would completely mess up the rich text editor. My
parser lays down a basic structure that pasted text/html can be
expected to follow, making it a lot easier to convert back and forth
from and to BBCode (and possibly Markdown). It also means markup
options are reasonably constrained so that they fit decently enough
with what F-Chat allows.

I've also added a DOM tree unzipper in order to reliably parse line
breaks. Newline characters will split the tree. I will write a
variation on this function that can be used to unzip the tree
arbitrarily, which will be useful for inserting other styling as well.

Overall, the Rich Text editor is well underway.

As a solution to F-Chat's rather restrictive use of BBCode, I will
hide unsupported tags in [i][/i] tags. For example, a heading would be
sent as such: [i=big][/i]Hello World[i=/big][/i]. However, the client
itself will still display this as [big]Hello World[/big]. It will also
be able to parse BBCode presented to it this way, even if other
clients cannot read it. This decision has been made primarily because
this is how slimCat functions; its custom tags will be presented to
the F-Chat webclient user unparsed. For the sake of compatibility,
therefore, these tags will be properly understood by Fermion. Tags
which Fermion doesn't understand should be silently removed from view
so the tags won't have any impact on the way the viewer experiences
the software. Presenting unparsed BBCode is something I see as
unprofessional. Instead, "invalid" BBCode will be ignored, much like
invalid CSS will be ignored by a browser. In essence, "invalid" BBCode
will simply function as "neutral" tags which have no effect. This is
also good for compatibility, as it will allow other clients to add new
tags without breaking Fermion's presentation.

There will still be an option *not* to parse invalid BBCode, mainly
for debugging purposes. This option will default to off, as, once
again, the default should be for Fermion to not break its presentation
when it encounters an unknown tag. Fermion will display all standard
tags, and there is no reason for a user to be made aware of something
that *doesn't* work if it doesn't otherwise affect the experience. At
most, it should be logged.

Also, I shouldn't rant in commit messages. Oops.