Improved BBCode parsing of 'hacked' tags.
'Hacked' tags can now, in addition, contain content themselves rather
than just be used as a vessel for carrying unsupported styling. It may
be worthwhile to disallow 'hacked' tags with content, so that if
content like '[i=left]foo[/i]' were received for some reason, it would
not be read as a 'hacked' tag by Fermion.

'[i=left]foo[/i]bar[i=/left]baz[/i]' parses more or less correctly
now. 'foo' and 'baz' are italicised, just like they would be in
F-Chat webclient. One potential problem is that the closing 'hacked'
tag puts 'baz' outside the left-aligned portion. I'll have to think on
whether this behaviour is desired.

'[i]foo[b]bar[/i]baz[/b]' still parses incorrectly. If I can get
nested tags to work as toggles it will show up correctly in Fermion,
but in F-Chat webclient the result is not as desired. My code appears
to create a functioning 'i' element from just the closing tag '[/i]'.
While this makes sure that the BBCode is always valid, this is not
what I want; if BBCode breaks on F-Chat webclient, Fermion compromises
the experience for vanilla users, so if BBCode is broken, I need to
show it as broken in Fermion, too, rather than sending it off as-is.
This will allow the user to fix incorrect BBCode for which no 'hacked'
tags exist to make it *not* break on F-Chat webclient's end.

Finally, certain F-Chat webclient tags don't allow sub-tags. Mainly
'[url]', '[sub]' and '[sup]', but also some F-Chat specific tags such
as '[icon]' and '[user]', for which sub-tags are nonsensical. This
functionality has not been implemented in the parser yet; these tags
can contain other tags in Fermion. Same as the above, if Fermion can't
send it in a way that won't break the webclient, it should consider
the BBCode in question broken.

Of course, if Fermion *receives* this invalid BBCode it may make an
attempt to correctly display it (the same tag nested inside itself
may function as a toggle, for example) but it will never *send* this
invalid BBCode.

There is also some Fermion-specific BBCode that won't make sense, such
as for example a '[big]' tag in a '[small]' tag, and so on. These
cases also need to be dealt with.