Fixed some issues with the BBCode parser.
There's a new function that cleans up the generated DOM, but it needs
more work. One of its tasks will be to check whether "invalid tags"
are actually tags (i.e. open + close are children of the same element)
or not. In the latter case, they will be converted to text nodes.

It might make sense to log messages in this format as well, rather
than how they are received/sent.