Made small tweaks to the BBCode parser, and wrote an auth stub.
The auth stub will be a way to safely store user credentials in
memory and later on disk without having them be accessible via JS, or
having them be readable as plain text from disk or RAM.

Also transitioning to the IIFE pattern to make true private variables,
and to reduce the possibility of name collisions, and prevent them
from silently breaking the application altogether. All returned values
are const, so they cannot be overwritten. Trying to change them will
result in an error.