Fixed the most glaring issues with the rewritten parser.
I'll still have to redo a bunch of tests, but the code does *feel*
cleaner. I'll have to reorder things a bit (place the classes next
to each other) but overall it's looking good.

I'll still have to wrestle with the markdown syntax. That's going to
be a big issue in and of itself. I can probably push that down the
line a bit while I work on other important features.