Wrote a number of tests for Awaiter and BBCode modules.
I know I'm supposed to write these *first* and then write my code,
but hey, at least I'm testing now. Coverage sucks for the BBCode
module, but it's hard to expose, since it needs awareness of the DOM.
I've installed jsdom to help with that, and also extracted the DOM-
aware function out of the module into the preload. Jest simply uses
template functions to help with testing. Kind of a hack job imo but
it works.